Move new validators to correct package
